DEV-505 move email verification to UserService
1. move email verification from registration service to user service
2. add new user status CREATED, change default status to CREATED
3. create new RegistrationServiceTest, move old one to
RegistrationServiceTestOld
