+++ service-registration/src/main/java/sg/ncl/service/registration/web/RegistrationController.java
@@ -75,9 +75,9 @@ public class RegistrationController {
         return registrationService.getDeterUid(id);
     }
 
-    @PutMapping(path = "/activation", params = {"uid", "key"})
+    /*@PutMapping(path = "/activation", params = {"uid", "key"})
     @ResponseStatus(HttpStatus.OK)
     private void activateAccount(@RequestParam("uid") final String uid, @RequestParam("key") final String key) {
         registrationService.activateAccount(uid, key);
-    }
+    }*/
 }
