+++ service-user/src/main/java/sg/ncl/service/user/domain/UserService.java
@@ -19,4 +19,6 @@ public interface UserService {
 
     void removeTeam(final String userId, final String teamId);
 
+    UserStatus verifyEmail(final String uid, final String email, final String key);
+
 }
